pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "systogony"
version = "0.0.1"
authors = [
    { name="David Chandler", email="contact@dechandler.io" }
]
description = "Simplified environment definition"
readme = "README.md"

license = "BSD-2-Clause"
license-files = ["LICENSE"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha"
]

requires-python = ">3.8"

dependencies = [
    "pyyaml",
    "declib @ git+https://github.com/dechandler/declib"
]


[project.urls]
"Code" = "https://github.com/dechandler/systogony"
"Documentation" = "https://github.com/dechandler/systogony/README.md"


[tool.setuptools]
packages = ["systogony"]
